Adjust `lib` e_shop psp split
