+++ config/splat.pspeu.stlib.yaml
@@ -76,9 +76,7 @@ segments:
       - [0x231B0, c, lib_psp/lang]
       - [0x23250, c, lib_psp/cutscene]
       - [0x25CD0, c, lib/st_debug]
-      - [0x25DB0, c, lib_psp/unk_25CD0]
-      - [0x290C8, c, lib_psp/e_shop]
-      - [0x2D788, c, lib_psp/2BC10]
+      - [0x25DB0, c, lib_psp/e_shop]
       - [0x35798, c, lib/e_lock_camera]
       - [0x35F60, c, lib/unk_3B200]
       - [0x36428, c, lib/e_candle_table]
@@ -122,9 +120,7 @@ segments:
       - [0x6D9C8, .rodata, lib_psp/1B7F0]
       - [0x6D9E8, .rodata, lib_psp/unk_1D9E8]
       - [0x6DA20, .rodata, lib_psp/cutscene]
-      - [0x6DAD8, .rodata, lib_psp/unk_25CD0]
-      - [0x6DB78, .rodata, lib_psp/e_shop]
-      - [0x6DBF8, .rodata, lib_psp/2BC10]
+      - [0x6DAD8, .rodata, lib_psp/e_shop]
       - [0x6DD20, .rodata, lib/unk_3B200]
       - [0x6DD80, .bss, lib/create_entity]
       - [0x6DDA0, .bss, lib/e_life_up]
