+++ config/splat.us.stlib.yaml
@@ -45,7 +45,7 @@ segments:
       - [0xAD8, .data, e_lock_camera]
       - [0xB58, data]
       - [0xDC0, .data, e_library_bg]
-      - [0xDF0, .data, e_shop]
+      - [0xE90, .data, e_shop]
       - [0x19BC, .data, cutscene]
       - [0x1A68, .data, e_candle_table]
       - [0x1A90, data]
@@ -99,7 +99,7 @@ segments:
       - [0x2DF98, c, e_lock_camera]
       - [0x2E4BC, c, first_c_file]
       - [0x2F280, c, e_library_bg]
-      - [0x2FA80, c, e_shop]
+      - [0x2FE0C, c, e_shop]
       - [0x395EC, c, cutscene]
       - [0x3AF60, c, e_candle_table]
       - [0x3B200, c, unk_3B200]
