+++ config/symbols.pspeu.stlib.txt
@@ -102,6 +102,7 @@ PrimDecreaseBrightness = 0x09251098;
 EntityThornweed = 0x09251168;
 EntityCorpseweed = 0x092515F8;
 EntityCorpseweedProjectile = 0x09252858;
+func_us_801AFA80 = 0x092537F8;
 func_us_801BB53C = 0x09254288;
 EntityBreakable = 0x092546D0;
 EntityChair = 0x09254DC0;
@@ -136,7 +137,25 @@ func_us_801B4080 = 0x09263628;
 func_us_801B40F0 = 0x092636F8;
 func_us_801B420C = 0x09263810;
 func_us_801B4830 = 0x092644E0;
+func_us_801B0C40 = 0x09264E20;
+func_us_801B4ED4 = 0x09265320;
+func_us_801B5068 = 0x09265608;
 func_us_801B0FBC = 0x09266028;
+func_us_801B1064 = 0x09266120;
+func_us_801B56E4 = 0x092663F8;
+func_us_801B5F18 = 0x09266E60;
+func_us_801B5F84 = 0x09266EE8;
+func_us_801B60C8 = 0x09267080;
+func_us_801B6124 = 0x092670F8;
+func_us_801B6324 = 0x09267400;
+func_us_801B6E20 = 0x09268360;
+func_us_801B6F30 = 0x09268528;
+func_us_801B7C94 = 0x092696F8;
+func_us_801B7DF8 = 0x092697D0;
+func_us_801B7D10 = 0x0926A128;
+func_us_801B8234 = 0x0926A238;
+func_us_801B8958 = 0x0926BD98;
+func_us_801B8A00 = 0x0926BE68;
 PlayerIsWithinHitbox = 0x0926CE18;
 EntityLockCamera = 0x0926CF18;
 func_us_801BB200 = 0x0926D5E0;
