--- src/st/lib_psp/2BC10.c
@@ -1,48 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../lib/lib.h"
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_09264E08);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_09264E20);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_09265320);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_09265608);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_us_801B0FBC);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_09266120);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_092663F8);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_09266E60);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_09266EE8);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_09267080);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_092670F8);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_09267400);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_09268360);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_09268528);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_092696F8);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_092697D0);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_09269FF0);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_0926A128);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_0926A238);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_0926ADD8);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_0926AED0);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_0926BD98);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/2BC10", func_psp_0926BE68);
