--- src/st/lib_psp/unk_25CD0.c
@@ -1,571 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../lib/lib.h"
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/unk_25CD0", func_psp_0925D430);
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/unk_25CD0", func_psp_0925D4D0);
-
-// This is probably EntityLibrarian, but I don't know for sure
-void func_us_801AFE0C(Entity* self) {
-    Tilemap* tilemap = &g_Tilemap;
-    Entity* player = &PLAYER;
-
-    switch (self->step) {
-    case 0:
-#ifdef VERSION_PSP
-        func_psp_0925D4D0();
-#endif
-        InitializeEntity(D_us_80180824);
-        if (player->posX.i.hi < 0x100) {
-// I expect these two sounds to be the same, but 0x202 has not yet been defined.
-// This leads me to think that the macro that has been defined for sfx 0x302
-// is only accurate for PSX and does not align with the sfx for PSP here.
-#ifdef VERSION_PSP
-            g_api.PlaySfx(0x302);
-#else
-            g_api.PlaySfx(0x202);
-#endif
-        }
-        if (g_CastleFlags[MET_LIBRARIAN]) {
-            self->step = 8;
-            break;
-        }
-#ifdef VERSION_PSP
-        g_Player.padSim = PAD_LEFT | PAD_SIM_UNK20000;
-        g_Player.D_80072EFC = 1;
-#endif
-        break;
-    case 1:
-        g_Entities[1].ext.entSlot1.unk0 = 1;
-        g_PauseAllowed = false;
-        g_unkGraphicsStruct.pauseEnemies = true;
-        g_Player.padSim = PAD_LEFT;
-        g_Player.D_80072EFC = 1;
-        if (g_Player.status & PLAYER_STATUS_BAT_FORM) {
-            g_Player.padSim = PAD_R1;
-        } else if (g_Player.status & PLAYER_STATUS_MIST_FORM) {
-#ifdef VERSION_PSP
-            g_Player.padSim = PAD_NONE;
-#else
-            g_Player.padSim = PAD_L1;
-#endif
-        } else if (g_Player.status & PLAYER_STATUS_WOLF_FORM) {
-#ifdef VERSION_PSP
-            g_Player.padSim = PAD_L1;
-#else
-            g_Player.padSim = PAD_R2;
-#endif
-        }
-        g_Player.D_80072EFC = 1;
-        SetStep(2);
-        break;
-    case 2:
-        if (player->posX.i.hi > 0xE8) {
-            if (g_Player.status & PLAYER_STATUS_TRANSFORM) {
-                g_Player.padSim = PAD_NONE;
-                if (g_Timer & 1) {
-                    if (g_Player.status & PLAYER_STATUS_BAT_FORM) {
-#ifdef VERSION_PSP
-                        g_Player.padSim = PAD_R1 | PAD_SIM_UNK20000;
-#else
-                        g_Player.padSim = PAD_R1;
-#endif
-                    } else if (g_Player.status & PLAYER_STATUS_MIST_FORM) {
-#ifdef VERSION_PSP
-                        g_Player.padSim = PAD_NONE | PAD_SIM_UNK20000;
-#else
-                        g_Player.padSim = PAD_L1;
-#endif
-                    } else if (g_Player.status & PLAYER_STATUS_WOLF_FORM) {
-#ifdef VERSION_PSP
-                        g_Player.padSim = PAD_L1 | PAD_SIM_UNK20000;
-#else
-                        g_Player.padSim = PAD_R2;
-#endif
-                    }
-                }
-            } else {
-                g_Player.padSim = PAD_LEFT;
-            }
-        } else {
-            g_CutsceneFlags |= 1;
-            g_Player.padSim = PAD_NONE;
-            player->posX.i.hi = 0xE8;
-            self->step++;
-        }
-        g_Player.D_80072EFC = 1;
-        break;
-    case 3:
-        if (g_CutsceneFlags & 0x40) {
-            if (player->posX.i.hi > 0x74) {
-                g_Entities[1].ext.entSlot1.unk0 = 1;
-                g_Player.padSim = PAD_LEFT;
-            } else {
-                player->posX.i.hi = 0x74;
-                g_Player.padSim = PAD_NONE;
-                self->step++;
-            }
-        } else {
-            player->posX.i.hi = 0xE8;
-        }
-        g_Player.D_80072EFC = 1;
-        break;
-    case 4:
-        g_Player.padSim = PAD_NONE | PAD_SIM_UNK20000;
-        g_Player.D_80072EFC = 1;
-        self->step++;
-        break;
-    case 5:
-        g_CastleFlags[MET_LIBRARIAN] = 1;
-        g_api.TimeAttackController(
-            TIMEATTACK_EVENT_MEET_MASTER_LIBRARIAN, TIMEATTACK_SET_RECORD);
-        g_Player.D_80072EFC = 1;
-        self->step++;
-        /* fallthrough */
-    case 6:
-        if (g_CutsceneFlags & 0x100) {
-            g_CutsceneFlags |= 0x2000;
-            self->step = 0x10;
-            break;
-        }
-        player->posX.i.hi = 0x74;
-        break;
-    case 8:
-        self->step++;
-        /* fallthrough */
-    case 9:
-        if (player->posX.i.hi > 0xFF) {
-            g_api.PlaySfx(CD_SOUND_COMMAND_7);
-            DestroyEntity(self);
-            break;
-        }
-        if (player->posX.i.hi < 0x75) {
-            switch (self->step_s) {
-            case 0:
-                g_Entities[1].ext.entSlot1.unk0 = 1;
-                g_PauseAllowed = false;
-                g_unkGraphicsStruct.pauseEnemies = true;
-                g_Player.padSim = PAD_NONE;
-                g_Player.D_80072EFC = 0x10;
-                self->step_s++;
-                g_CutsceneFlags |= 1;
-                break;
-            case 1:
-                if (g_Player.status & PLAYER_STATUS_TRANSFORM) {
-                    g_Player.padSim = PAD_NONE;
-                    if (g_Timer & 1) {
-                        if (g_Player.status & PLAYER_STATUS_BAT_FORM) {
-                            g_Player.padSim = PAD_R1;
-                        } else if (g_Player.status & PLAYER_STATUS_MIST_FORM) {
-#ifdef VERSION_PSP
-                            g_Player.padSim = PAD_NONE;
-#else
-                            g_Player.padSim = PAD_L1;
-#endif
-                        } else if (g_Player.status & PLAYER_STATUS_WOLF_FORM) {
-#ifdef VERSION_PSP
-                            g_Player.padSim = PAD_L1;
-#else
-                            g_Player.padSim = PAD_R2;
-#endif
-                        }
-                    }
-                } else {
-                    g_Player.padSim = PAD_LEFT;
-                    self->step_s++;
-                }
-                g_Player.D_80072EFC = 1;
-                break;
-            case 2:
-                g_Player.padSim = PAD_NONE;
-                g_Player.D_80072EFC = 0x80;
-                SetStep(10);
-                break;
-            }
-            player->posX.i.hi = 0x74;
-        }
-        break;
-    case 10:
-        if (!g_Player.D_80072EFC && (g_Player.pl_vram_flag & 1)) {
-            g_Player.padSim = PAD_NONE | PAD_SIM_UNK20000;
-            g_Player.D_80072EFC = 1;
-            self->step++;
-        }
-        player->posX.i.hi = 0x74;
-        break;
-    case 11:
-        g_Player.padSim = PAD_NONE | PAD_SIM_UNK20000;
-        g_Player.D_80072EFC = 1;
-        if (g_CutsceneFlags & 0x100) {
-            g_CutsceneFlags |= 0x2000;
-            self->step = 0x10;
-        }
-        break;
-    case 16:
-#ifdef VERSION_PSP
-        g_PauseAllowed = false;
-#endif
-        g_Player.D_80072EFC = 0x20;
-        g_Player.padSim = PAD_RIGHT;
-        D_80097928 = 1;
-        self->step++;
-        break;
-    case 17:
-#ifdef VERSION_PSP
-        g_PauseAllowed = false;
-#endif
-        if (!g_Player.D_80072EFC) {
-#ifdef VERSION_PSP
-            g_PauseAllowed = true;
-#endif
-            SetStep(9);
-        }
-        break;
-    }
-}
-
-extern u8 D_us_801811FC[];
-extern u8 D_us_80181204[];
-extern u8 D_us_80181210[];
-extern u8 D_us_8018121C[];
-extern u8 D_us_80181234[];
-extern u8 D_us_80181244[];
-extern u8 D_us_801812C0[];
-extern u8 D_us_801812C8[];
-extern u8 D_us_801812D0[];
-extern u8* D_us_801812D8[];
-
-void EntityLibrarianChair(Entity* self) {
-    Entity* newEnt;
-    Entity* player = &PLAYER;
-    Tilemap* tilemap = &g_Tilemap;
-
-    if (self->step && (self->step < 11) && (g_SkipCutscene != 0) &&
-        (g_IsCutsceneDone != 0)) {
-        self->step = 11;
-        self->animCurFrame = 2;
-    }
-    if (player->posX.i.hi < 0x38) {
-        self->zPriority = 0xC0;
-        if (g_Player.status & PLAYER_STATUS_TRANSFORM) {
-            GetPlayerCollisionWith(self, 16, 12, 2);
-            // If the chair is not in step 16, and the player is high enough,
-            // and moving upward, trigger a hit.
-        } else if (self->step < 16 &&
-                   ((player->posY.i.hi + tilemap->scrollY.i.hi) < 201) &&
-                   ((*((f32*)&player->velocityY)).i.hi < 0)) {
-            SetStep(16);
-            if (PLAYER.step == Player_HighJump) {
-                g_Player.unk4A = 0x1C;
-                if (PLAYER.step_s == 0) {
-                    PLAYER.step_s = 1;
-                    PLAYER.velocityY = FIX(-8);
-                }
-            }
-            // Play chair jumping sound
-            g_api.PlaySfx(SFX_QUICK_STUTTER_EXPLODE_B);
-            self->ext.libraryChair.totalHits++;
-            self->ext.libraryChair.consecutiveHits++;
-
-            // At first hit, give Life Max Up.
-            if (!(g_CastleFlags[LIBRARIAN_DROPS] & 1)) {
-                newEnt = AllocEntity(&g_Entities[160], &g_Entities[192]);
-                if (newEnt != NULL) {
-                    CreateEntityFromCurrentEntity(E_PRIZE_DROP, newEnt);
-                    newEnt->params = 23;
-                    g_CastleFlags[LIBRARIAN_DROPS] |= 1;
-                }
-            }
-            // Getting Axe Lord Armor requires hitting librarian 64 times.
-            if (!(g_CastleFlags[LIBRARIAN_DROPS] & 2) &&
-                (self->ext.libraryChair.totalHits >= 64)) {
-                newEnt = AllocEntity(&g_Entities[160], &g_Entities[192]);
-                if (newEnt != NULL) {
-                    CreateEntityFromCurrentEntity(E_EQUIP_ITEM_DROP, newEnt);
-                    newEnt->params = NUM_HAND_ITEMS + ITEM_AXE_LORD_ARMOR;
-                    g_CastleFlags[LIBRARIAN_DROPS] |= 2;
-                }
-            }
-            // Ring of Arcana requires 16 hits, without touching ground
-            if (!(g_CastleFlags[LIBRARIAN_DROPS] & 4) &&
-                (self->ext.libraryChair.consecutiveHits >= 16)) {
-                newEnt = AllocEntity(&g_Entities[160], &g_Entities[192]);
-                if (newEnt != NULL) {
-                    CreateEntityFromCurrentEntity(E_EQUIP_ITEM_DROP, newEnt);
-                    newEnt->params = NUM_HAND_ITEMS + ITEM_RING_OF_ARCANA;
-                    g_CastleFlags[LIBRARIAN_DROPS] |= 4;
-                }
-            }
-            // Dracula Tunic requires 24 hits, and inverted castle must be
-            // unlocked.
-            if (!(g_CastleFlags[LIBRARIAN_DROPS] & 8) &&
-                (self->ext.libraryChair.consecutiveHits >= 24) &&
-                (g_CastleFlags[INVERTED_CASTLE_UNLOCKED])) {
-                newEnt = AllocEntity(&g_Entities[160], &g_Entities[192]);
-                if (newEnt != NULL) {
-                    CreateEntityFromCurrentEntity(E_EQUIP_ITEM_DROP, newEnt);
-                    newEnt->params = NUM_HAND_ITEMS + ITEM_DRACULA_TUNIC;
-                    g_CastleFlags[LIBRARIAN_DROPS] |= 8;
-                }
-            }
-        }
-    } else {
-        self->zPriority = 0x80;
-    }
-    // If the player touches the ground, reset the frames airborne.
-    if (g_Player.pl_vram_flag & 1) {
-        self->ext.libraryChair.consecutiveHits = 0;
-    }
-    switch (self->step) {
-    case 0:
-        InitializeEntity(g_EInitCommon);
-        self->zPriority = 0x80;
-        self->animSet = ANIMSET_OVL(11);
-        self->animCurFrame = 2;
-        self->palette = 0x210;
-        self->unk5A = 0x48;
-        self->ext.libraryChair.debugAnimID = 0;
-        self->ext.libraryChair.timer = 0x20;
-        self->flags &= ~FLAG_DESTROY_IF_OUT_OF_CAMERA;
-        if (g_PlayableCharacter != PLAYER_ALUCARD) {
-            self->facingLeft = 1;
-            self->posX.i.hi -= 8;
-        }
-        g_CutsceneFlags = 0;
-        g_IsCutsceneDone = 0;
-        g_SkipCutscene = 0;
-        break;
-    case 1:
-        if (g_CutsceneFlags & 2) {
-            if (--self->ext.libraryChair.timer) {
-                SetStep(2);
-            }
-        } else {
-            if (g_CutsceneFlags & 0x80) {
-                SetStep(9);
-            }
-        }
-        break;
-    case 2:
-        if (AnimateEntity(D_us_801811FC, self) == 0) {
-            SetStep(3);
-        }
-        break;
-    case 3:
-        AnimateEntity(D_us_80181204, self);
-        if (g_CutsceneFlags & 4) {
-            SetStep(4);
-        }
-        break;
-    case 4:
-        AnimateEntity(D_us_80181210, self);
-        if (g_CutsceneFlags & 8) {
-            SetStep(5);
-        }
-        break;
-    case 5:
-        AnimateEntity(D_us_8018121C, self);
-        if (g_CutsceneFlags & 0x10) {
-            SetStep(6);
-        }
-        break;
-    case 6:
-        AnimateEntity(D_us_80181234, self);
-        if (g_CutsceneFlags & 0x20) {
-            SetStep(7);
-        }
-        break;
-    case 7:
-        if (AnimateEntity(D_us_80181244, self) == 0) {
-            SetStep(11);
-        }
-        break;
-    case 9:
-        if (AnimateEntity(D_us_801811FC, self) == 0) {
-            SetStep(10);
-        }
-        break;
-    case 10:
-        if (AnimateEntity(D_us_801812C0, self) == 0) {
-            SetStep(11);
-        }
-        break;
-    case 11:
-        if (g_CutsceneFlags & 0x800) {
-            SetStep(12);
-            self->ext.libraryChair.timer = 0x30;
-        }
-        if (g_CutsceneFlags & 0x1000) {
-            SetStep(13);
-        }
-        break;
-    case 12:
-        AnimateEntity(D_us_801812C8, self);
-        if (!--self->ext.libraryChair.timer) {
-            g_CutsceneFlags &= ~0x800;
-            self->animCurFrame = 2;
-            SetStep(11);
-        }
-        if (g_CutsceneFlags & 0x1000) {
-            self->animCurFrame = 2;
-            SetStep(13);
-        }
-        break;
-    case 13:
-        break;
-    case 16:
-        AnimateEntity(D_us_801812D0, self);
-        if (g_Player.status & PLAYER_STATUS_TRANSFORM) {
-            self->velocityY += FIX(3.0 / 8);
-            if (self->velocityY > FIX(7)) {
-                self->velocityY = FIX(7);
-            }
-            MoveEntity();
-            GetPlayerCollisionWith(self, 16, 12, 0x12);
-        } else {
-            if (PLAYER.step == Player_HighJump && PLAYER.step_s == 0) {
-                PLAYER.step_s = 1;
-                PLAYER.velocityY = FIX(-8);
-            }
-            self->posY.i.hi = player->posY.i.hi - 0x1C;
-            self->velocityY = player->velocityY;
-        }
-        if ((self->posY.i.hi + tilemap->scrollY.i.hi) >= 172) {
-            self->posY.i.hi = 172 - tilemap->scrollY.i.hi;
-            self->animCurFrame = 2;
-            SetStep(1);
-        }
-        break;
-    case 255:
-        if (g_pads[0].tapped & PAD_UP) {
-            if (self->ext.libraryChair.debugAnimID) {
-                self->ext.libraryChair.debugAnimID--;
-                self->animFrameIdx = 0;
-                self->animFrameDuration = 0;
-            }
-        } else if (g_pads[0].tapped & PAD_DOWN) {
-            if (self->ext.libraryChair.debugAnimID != 0x10) {
-                self->ext.libraryChair.debugAnimID++;
-                self->animFrameIdx = 0;
-                self->animFrameDuration = 0;
-            }
-        }
-        AnimateEntity(D_us_801812D8[self->ext.libraryChair.debugAnimID], self);
-        break;
-    }
-}
-
-void func_us_801B11A0(s16 x, s16 y, u16 w, u16 h) {
-    RECT rect;
-
-    rect.x = x / 4;
-    rect.y = y;
-    rect.w = w / 4 + 1;
-    rect.h = h + 1;
-    ClearImage(&rect, 0, 0, 0);
-}
-
-void func_us_801B1200(Primitive* prim, Primitive* otherPrim) {
-    prim->x0 = otherPrim->x0;
-    prim->y0 = otherPrim->y0;
-    prim->x1 = otherPrim->x1;
-    prim->y1 = otherPrim->y1;
-    prim->drawMode = DRAW_DEFAULT;
-
-    prim = prim->next;
-    prim->x0 = otherPrim->x1;
-    prim->y0 = otherPrim->y1;
-    prim->x1 = otherPrim->x3;
-    prim->y1 = otherPrim->y3;
-    prim->drawMode = DRAW_DEFAULT;
-
-    prim = prim->next;
-    prim->x0 = otherPrim->x0;
-    prim->y0 = otherPrim->y0;
-    prim->x1 = otherPrim->x2;
-    prim->y1 = otherPrim->y2;
-    prim->drawMode = DRAW_DEFAULT;
-
-    prim = prim->next;
-    prim->x0 = otherPrim->x2;
-    prim->y0 = otherPrim->y2;
-    prim->x1 = otherPrim->x3;
-    prim->y1 = otherPrim->y3;
-    prim->drawMode = DRAW_DEFAULT;
-}
-
-void func_us_801B12D0(Entity* self, u16 arg1) {
-    Primitive* prim2;
-    Primitive* prim;
-    s16 s_4;
-    s16 posY;
-    s16 posX;
-    s32 i;
-    s16 s_3;
-    s16 temp_v1;
-    s16 var_s4;
-
-    prim = &g_PrimBuf[g_CurrentEntity->primIndex];
-    prim = prim->next;
-    prim2 = prim;
-    for (i = 0; i < 4; i++) {
-        prim = prim->next;
-    }
-    i = 0;
-    s_3 = g_CurrentEntity->ext.et_801B15C0.unk86;
-    s_4 = g_CurrentEntity->ext.et_801B15C0.unk80 * 0x10;
-#ifdef VERSION_PSP
-    if (arg1) {
-        posY = (g_CurrentEntity->ext.et_801B15C0.unk7C * ((s_3 + 0x18) - s_4)) /
-               24;
-        posX = (g_CurrentEntity->ext.et_801B15C0.unk7C * -0x88) / 24;
-#else
-    if (arg1 & 1) {
-        posY = (g_CurrentEntity->ext.et_801B15C0.unk7C * ((s_3 + 0x18) - s_4)) /
-               24;
-        posX = (g_CurrentEntity->ext.et_801B15C0.unk7C * -0x98) / 24;
-#endif
-    } else {
-        posY = (g_CurrentEntity->ext.et_801B15C0.unk7C * ((s_3 + 0x40) - s_4)) /
-               24;
-        posX = 0;
-    }
-#ifdef VERSION_PSP
-    for (; prim != NULL; i++) {
-        prim->x0 = prim->x2 = posX + 0x94;
-        prim->x1 = prim->x3 = posX + 0xEC;
-#else
-    if (arg1 & 0x10) {
-        s32 temp = g_CurrentEntity->ext.et_801B15C0.unk7C;
-        var_s4 = temp / 3;
-    } else {
-        var_s4 = 0;
-    }
-    for (; prim != NULL; i++) {
-        prim->x0 = prim->x2 = posX + 0x96 + var_s4;
-        prim->x1 = prim->x3 = posX + 0xEA - var_s4;
-        if (i != 0) {
-            prim->u0 = prim->u2 = var_s4;
-            prim->u1 = prim->u3 = 0x54 - var_s4;
-        }
-#endif
-        temp_v1 = (g_CurrentEntity->ext.et_801B15C0.unk7C * (s_4 + 2)) / 24;
-        if (i != 0) {
-            prim->v0 = prim->v1 = temp_v1;
-        }
-        prim->y0 = prim->y1 = (0x60 - s_3) + temp_v1 + posY;
-        temp_v1 = (g_CurrentEntity->ext.et_801B15C0.unk7C *
-                   (s_3 * 2 - (s_4 + 0x15))) /
-                  24;
-        if (i != 0) {
-            prim->v2 = prim->v3 = s_3 * 2 - temp_v1;
-        }
-        prim->y2 = prim->y3 = (s_3 + 0x60) - temp_v1 + posY;
-        if (i == 0) {
-            func_us_801B1200(prim2, prim);
-        }
-        prim = prim->next;
-    }
-}
-
-INCLUDE_ASM("st/lib_psp/psp/lib_psp/unk_25CD0", func_us_801B15C0);
